Add a new command for generic HTTP requests.
git-svn-id: https://svn.filezilla-project.org/svn/FileZilla3/trunk@8999 ce1b9ea0-41a8-11dc-9b12-cb80f3b22977

+++ src/engine/notification.cpp
@@ -60,7 +60,7 @@ int CHostKeyNotification::GetPort() const
 	return m_port;
 }
 
-CDataNotification::CDataNotification(char* pData, int len)
+CDataNotification::CDataNotification(char* pData, size_t len)
 	: m_pData(pData), m_len(len)
 {
 }
@@ -70,7 +70,7 @@ CDataNotification::~CDataNotification()
 	delete [] m_pData;
 }
 
-char* CDataNotification::Detach(int& len)
+char* CDataNotification::Detach(size_t& len)
 {
 	len = m_len;
 	char* pData = m_pData;
